+++ rust/src/libponyc/codegen/genname.rs
@@ -635,7 +635,7 @@ unsafe extern "C" fn type_append(
                 type_0,
                 (::core::mem::size_of::<[*mut *mut ast_t; 6]>() as libc::c_ulong)
                     .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
-                    .wrapping_sub(1 as libc::c_int as libc::c_ulong),
+                    .wrapping_sub(1),
                 children.as_mut_ptr(),
             );
             let mut def: *mut ast_t = ast_data(type_0) as *mut ast_t;
