+++ rust/src/libponyc/type/sanitise.rs
@@ -632,7 +632,7 @@ unsafe extern "C" fn collect_type_param(
         orig_param,
         (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
             .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
-            .wrapping_sub(1 as libc::c_int as libc::c_ulong),
+            .wrapping_sub(1),
         children.as_mut_ptr(),
     );
     let mut name: *const libc::c_char = ast_name(id);
